Fix issue by splitting tests into headless and non-headless parts, and then skip the non-headless part on GitHub Actions.
