+++ tox.ini
@@ -18,6 +18,7 @@ commands =
         --ignore=test/browser/combo/cookie_banner_test.py \
         --ignore=test/browser/combo/login_test.py \
         --ignore=test/browser/combo/search_test.py \
+        --ignore=test/browser/get/screen_size_non_headless_test.py \
         --ignore=test/browser/window/get/position_1_test.py \
         --ignore=test/browser/window/get/size_1_test.py \
         --ignore=test/browser/window/set/position_2_test.py \
